chore: fix regulation modal styles
