+++ packages/tradershub/src/components/RegulationTableContent/index.ts
@@ -1,0 +1,1 @@
+export { default as RegulationTableContent } from './RegulationTableContent';
